+++ packages/picasso-lab/src/QuoteMark/styles.ts
@@ -1,0 +1,8 @@
+import { Theme, createStyles } from '@material-ui/core'
+
+export default ({ palette }: Theme) =>
+  createStyles({
+    root: {
+      fill: palette.primary.main
+    }
+  })