+++ packages/picasso/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.5.0](https://github.com/toptal/picasso/compare/@toptal/picasso@5.4.9...@toptal/picasso@5.5.0) (2021-01-25)
+
+
+### Features
+
+* **Typography:** add option to inherit weight ([#1862](https://github.com/toptal/picasso/issues/1862)) ([146d3a4](https://github.com/toptal/picasso/commit/146d3a44926ed515c5cc284f3dbd2d459efd39bd))
+
+
+
+
+
+## [5.4.9](https://github.com/toptal/picasso/compare/@toptal/picasso@5.4.8...@toptal/picasso@5.4.9) (2021-01-21)
+
+**Note:** Version bump only for package @toptal/picasso
+
+
+
+
+
+## [5.4.8](https://github.com/toptal/picasso/compare/@toptal/picasso@5.4.7...@toptal/picasso@5.4.8) (2021-01-21)
+
+
+### Bug Fixes
+
+* **Input:** fix autocomplete edges non-clickable ([#1855](https://github.com/toptal/picasso/issues/1855)) ([772607a](https://github.com/toptal/picasso/commit/772607a0ac6b8f66359b784f5834370a72bc5d83))
+
+
+
+
+
 ## [5.4.7](https://github.com/toptal/picasso/compare/@toptal/picasso@5.4.6...@toptal/picasso@5.4.7) (2021-01-20)